Refactor StorageManager to hide concrete implementations
Refactor StorageManager into its own subpackage. Additionally, move
FilesystemStorageManager to its own private module. Expose a create
function in storagemanager that returns the appropriate concrete
StorageManager class based off the arguments given.
